+++ src/main/java/br/dev/andrestamatto/identityhub/infrastructure/messaging/template/email/UserVerificationCodeEmailTemplate.java
@@ -1,4 +1,4 @@
-package br.dev.andrestamatto.identityhub.infrastructure.messaging.templates;
+package br.dev.andrestamatto.identityhub.infrastructure.messaging.template.email;
 
 
 import br.dev.andrestamatto.identityhub.application.ports.output.messaging.NotificationMessage;
@@ -7,23 +7,23 @@ import org.thymeleaf.context.Context;
 import org.thymeleaf.spring6.SpringTemplateEngine;
 
 /**
- * Thymeleaf renderer for the welcome email sent after account confirmation.
+ * Thymeleaf renderer for the user verification-code email.
  * It maps NotificationMessage details into the HTML template stored in resources/templates/emails.
  */
-public class UserWelcomeEmailTemplate implements EmailTemplate {
+public class UserVerificationCodeEmailTemplate implements EmailTemplate {
 
     private static final String EMAIL_TEMPLATES_PATH = "emails/";
-    private final static String USER_WELCOME_TEMPLATE = EMAIL_TEMPLATES_PATH + "user-welcome-template";
+    private final static String USER_VERIFICATION_CODE_TEMPLATE = EMAIL_TEMPLATES_PATH + "user-verification-code-template";
 
     private final SpringTemplateEngine springTemplateEngine;
 
-    public UserWelcomeEmailTemplate(SpringTemplateEngine springTemplateEngine) {
+    public UserVerificationCodeEmailTemplate(SpringTemplateEngine springTemplateEngine) {
         this.springTemplateEngine = springTemplateEngine;
     }
 
     @Override
     public EmailMessageTemplate supports() {
-        return EmailMessageTemplate.EMAIL_USER_SUCCESSFULLY_ACTIVATED;
+        return EmailMessageTemplate.EMAIL_USER_VERIFICATION_CODE;
     }
 
     @Override
@@ -33,7 +33,7 @@ public class UserWelcomeEmailTemplate implements EmailTemplate {
         context.setVariable("username", notificationMessage.recipient());
         notificationMessage.details().forEach(context::setVariable);
 
-        return springTemplateEngine.process(USER_WELCOME_TEMPLATE, context);
+        return springTemplateEngine.process(USER_VERIFICATION_CODE_TEMPLATE, context);
     }
 
 }