+++ src/test/java/br/dev/andrestamatto/identityhub/infrastructure/messaging/template/email/UserWelcomeEmailTemplateTest.java
@@ -1,10 +1,11 @@
-package br.dev.andrestamatto.identityhub.infrastructure.messaging.templates;
+package br.dev.andrestamatto.identityhub.infrastructure.messaging.template.email;
 
 import br.dev.andrestamatto.identityhub.application.ports.output.messaging.NotificationMessage;
 import br.dev.andrestamatto.identityhub.application.ports.output.messaging.channels.NotificationChannels;
 import br.dev.andrestamatto.identityhub.application.ports.output.messaging.templates.EmailMessageTemplate;
 import br.dev.andrestamatto.identityhub.application.ports.output.messaging.templates.MessageTemplates;
 import br.dev.andrestamatto.identityhub.application.ports.output.messaging.templates.SmsMessageTemplate;
+import br.dev.andrestamatto.identityhub.infrastructure.messaging.template.email.UserWelcomeEmailTemplate;
 import br.dev.andrestamatto.identityhub.support.UserTestData;
 import org.junit.jupiter.api.Test;
 import org.thymeleaf.spring6.SpringTemplateEngine;